removeBg_api_python/requirements.txt
flask
flask_cors
pillow
rembg
numpy

